Fixed another bug in evaluation method for cost functional.
The problem was that for the forward and backward solve I use explicit Euler, but in the evaluation of the cost functional the trapezoidal rule was used for the time integration of y. This causes a problem with consistency which should be fixed now. You can observe that the finite difference gradient and the adjoint gradient are the same now.

Also there was another issue that in the solution of the adjoint system a wrong value of y was used.